requirements.txt
streamlit
google-genai
PyPDF2
python-docx
google-generativeai
